Make it easier to make multiple tests
