+++ app/api.ts
@@ -1,4 +1,4 @@
-import { z } from "zod";
+import { JoinFormInput, JoinFormResponse, NNAssignFormInput, NNAssignFormResponse } from "@/app/io";
 
 if (process.env.NEXT_PUBLIC_MESHDB_URL === undefined) {
   throw new Error('Expected API url environment variable');
@@ -9,78 +9,6 @@ if (process.env.NEXT_PUBLIC_MESHDB_URL === undefined) {
 //}
 
 const API_BASE = new URL(process.env.NEXT_PUBLIC_MESHDB_URL as string + "/api/v1/");
-
-export const JoinFormInput = z.object({
-  first_name: z.string(),
-  last_name: z.string(),
-  email: z.string(),
-  phone: z.string(),
-  street_address: z.string(),
-  apartment: z.string(),
-  city: z.string(),
-  state: z.string(),
-  zip: z.number(),
-  roof_access: z.boolean(),
-  referral: z.string(),
-  ncl: z.boolean(),
-})
-export type JoinFormInput = z.infer<typeof JoinFormInput>
-
-export const JoinFormResponse = z.object({
-  message: z.string().optional(),
-  building_id: z.number(),
-  member_id: z.number(),
-  install_number: z.number(),
-  member_exists: z.boolean(),
-})
-export type JoinFormResponse = z.infer<typeof JoinFormResponse>
-
-export const NNAssignFormInput = z.object({
-  install_number: z.number(),
-  password: z.string(), // TODO: Salt/hash/whatever this
-})
-export type NNAssignFormInput = z.infer<typeof NNAssignFormInput>
-
-export const NNAssignFormResponse = z.object({
-  message: z.string().optional(),
-  building_id: z.number(),
-  install_number: z.number(),
-  network_number: z.number(),
-  created: z.boolean(),
-})
-export type NNAssignFormResponse = z.infer<typeof NNAssignFormResponse>
-
-export const QueryFormInput = z.object({
-  //route: z.string(),
-  legacy: z.string().optional(),
-  query_type: z.string(),
-  data: z.string(),
-  password: z.string(), // TODO: Salt/hash/whatever this
-})
-export type QueryFormInput = z.infer<typeof QueryFormInput>
-
-export const QueryFormResponse = z.object({
-  count: z.number(),
-  next: z.string().nullable(),
-  previous: z.string().nullable(),
-  results: z.array(z.object({
-    install_number: z.number(),
-    street_address: z.string().nullable(),
-    unit: z.string(),
-    city: z.string(),
-    state: z.string(),
-    zip_code: z.string(),
-    name: z.string(),
-    phone_number: z.string().nullable(),
-    primary_email_address: z.string().nullable(),
-    stripe_email_address: z.string().nullable(),
-    additional_email_addresses: z.array(z.string()).nullable(),
-    notes: z.string(),
-    network_number: z.number().nullable(),
-    status: z.string(),
-  }))
-})
-export type QueryFormResponse = z.infer<typeof QueryFormResponse>
 
 const get = async <S extends z.Schema>(url: string, schema: S, auth?: string, nextOptions?: NextFetchRequestConfig): Promise<ReturnType<S['parse']>> => {
   const res = await fetch(new URL(url, API_BASE), {
