+++ tests/join_form.spec.ts
@@ -1,7 +1,14 @@
-import { test, expect } from '@playwright/test';
+import { JoinFormInput } from '@/app/io';
+import { test, expect, Page } from '@playwright/test';
+
+// Integration tests for the Join Form.
+//
+// These tests can hit either a self-hosted dev instance of MeshDB
+// (See https://github.com/nycmeshnet/meshdb) or it can hit the beta
+// env.
 
 // TODO (wdn): I want to test
-// - Full cooperation with real address
+// - Full cooperation with real address X
 // - Missing address
 // - Missing phone
 // - Missing email
@@ -11,6 +18,56 @@ import { test, expect } from '@playwright/test';
 // - Bad email
 // Can we mirror what meshdb does?
 
+async function fillOutJoinForm(page: Page, sampleData: JoinFormInput) {
+  // Set up some sample data
+
+  // Personal info
+  await page
+    .getByPlaceholder('First Name')
+    .fill(sampleData.first_name);
+  await page
+    .getByPlaceholder('Last Name')
+    .fill(sampleData.last_name);
+  await page
+    .getByPlaceholder('Email Address')
+    .fill(sampleData.email);
+  await page
+    .getByPlaceholder('Phone Number')
+    .fill(sampleData.phone);
+
+  // Address Info
+  await page
+    .getByPlaceholder('Street Address')
+    .fill(sampleData.street_address);
+
+  await page
+    .getByPlaceholder('Unit #')
+    .fill(sampleData.apartment);
+
+  await page
+    .getByPlaceholder('City')
+    .fill(sampleData.city);
+
+  await page
+    .getByPlaceholder('Zip Code')
+    .fill(sampleData.zip.toString());
+
+  // How did you hear about us?
+  await page
+    .getByPlaceholder('How did you hear about us?')
+    .fill(sampleData.referral);
+
+  // Agree to the NCL
+  if (sampleData.ncl) {
+    await page.locator("[name='ncl']").check();
+  }
+
+  // Roof Access
+  if (sampleData.roof_access) {
+    await page.locator("[name='roof_access']").check();
+  }
+}
+
 // This tests the actual form.
 test('happy join form', async ({ page }) => {
   test.setTimeout(10000)
@@ -19,45 +76,23 @@ test('happy join form', async ({ page }) => {
   // Is the page title correct?
   await expect(page).toHaveTitle(/Join Our Community Network!/);
 
-  // Set up some sample data
-  await page
-    .getByPlaceholder('First Name')
-    .fill('Jon');
-  await page
-    .getByPlaceholder('Last Name')
-    .fill('Smith');
-  await page
-    .getByPlaceholder('Email Address')
-    .fill('js@gmail.com');
+  const sampleData: JoinFormInput = JoinFormInput.parse({
+    first_name: "Jon",
+    last_name: "Smith",
+    email: "js@gmail.com",
+    phone: "585-475-2411",
+    street_address: "876 Bergen St",
+    apartment: "7",
+    city: "Brooklyn",
+    state: "NY",
+    zip: 11238,
+    roof_access: true,
+    referral: "I googled it.",
+    ncl: true,
+  });
 
-  await page
-    .getByPlaceholder('Phone Number')
-    .fill('585-475-2411');
-
-  // Address Info
-  await page
-    .getByPlaceholder('Street Address')
-    .fill('876 Bergen St');
-
-  await page
-    .getByPlaceholder('Unit #')
-    .fill('7');
-
-  await page
-    .getByPlaceholder('City')
-    .fill('Brooklyn');
-
-  await page
-    .getByPlaceholder('Zip Code')
-    .fill('11238');
-
-  // How did you hear about us?
-  await page
-    .getByPlaceholder('How did you hear about us?')
-    .fill('I googled it.');
-
-  // Agree to the NCL
-  await page.locator("[name='ncl']").check();
+  // Set up sample data.
+  await fillOutJoinForm(page, sampleData);
   
   // Submit the join form
   await page
